Add config title to modal
