+++ client/src/pages/MainPage/Header/index.jsx
@@ -3,22 +3,14 @@ import { AiOutlineMenu } from 'react-icons/ai'
 import Tooltip from '../../../components/Tooltip'
 import { useState, useRef, useEffect } from 'react'
 import { GoGear } from 'react-icons/go'
-import { AnimatePresence } from 'framer-motion'
-import Modal from '../../../components/Modal'
-import usePlayer from '../../../hooks/usePlayer'
+import ConfigModal from './ConfigModal'
 
 export default function Header() {
     const [openMenu, setOpenMenu] = useState(false)
     const [openConfigModal, setOpenConfigModal] = useState(false)
     const [openConfirmationModal, setOpenConfirmationModal] = useState(false)
-    const { setPlayerData, defaultData }= usePlayer()
 
     const menuRef = useRef()
-
-    const restoreGameData = () => {
-        setPlayerData(defaultData)
-        setOpenConfirmationModal(false)
-    }
 
     useEffect(() => {
         let handler = e => {
@@ -59,65 +51,15 @@ export default function Header() {
                 </C.MenuContainer>
             </C.Header>
 
-            <AnimatePresence
-                initial={false}
-                mode='wait'
-                onExitComplete={() => null}
-            >
-                {openConfigModal &&
-                    <Modal
-                        handleClose={() => setOpenConfigModal(false)}
-                        size='small'
-                        animation='fadeIn'>
+            {/* Menu Nav Modals */}
 
-                        <C.ModalContent>
-                            <C.RestoreDataButton
-                                whileHover={{ scale: 1.05 }}
-                                whileTap={{ scale: 0.95 }}
-                                onClick={() => setOpenConfirmationModal(true)}
-                            >
-                                Restore Game Data
-                            </C.RestoreDataButton>
-                        </C.ModalContent>
-                    </Modal>
-                }
-            </AnimatePresence>
+            <ConfigModal
+                openConfigModal={openConfigModal}
+                setOpenConfigModal={setOpenConfigModal}
+                openConfirmationModal={openConfirmationModal}
+                setOpenConfirmationModal={setOpenConfirmationModal}
+            />
 
-            <AnimatePresence
-                initial={false}
-                mode='wait'
-                onExitComplete={() => null}
-            >
-                {openConfigModal && openConfirmationModal &&
-                    <Modal
-                        handleClose={() => setOpenConfirmationModal(false)}
-                        size='tiny'
-                        animation='fadeIn'>
-
-                        <C.ModalContent>
-                            <p className='font-medium text-center'>Are you sure you want to restore the game to its default data settings?</p>
-                            <div className='flex items-center justify-center gap-8 w-full p-5'>
-                                <C.ConfirmationButton
-                                    whileHover={{ scale: 1.1 }}
-                                    whileTap={{ scale: 0.9 }}
-                                    onClick={() => setOpenConfirmationModal(false)}
-                                    action='close'
-                                >
-                                    Close
-                                </C.ConfirmationButton>
-
-                                <C.ConfirmationButton
-                                    whileHover={{ scale: 1.1 }}
-                                    whileTap={{ scale: 0.9 }}
-                                    onClick={restoreGameData}
-                                >
-                                    Confirm
-                                </C.ConfirmationButton>
-                            </div>
-                        </C.ModalContent>
-                    </Modal>
-                }
-            </AnimatePresence>
         </>
     )
 }
